change to integer

+++ lib/bmb/product.ex
@@ -6,7 +6,7 @@ defmodule Bmb.Product do
   schema "products" do
     field :name, :string
     field :display_name, :string
-    field :price, :decimal
+    field :price, :integer
     field :meta_title, :string
     field :meta_description, :string
     field :meta_keyword, :string
